Create a separate handler to avoid import cycles.
